+++ src/main/java/roomescape/HomeController.java
@@ -1,13 +1,13 @@
-package roomescape.reservation;
+package roomescape;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class ViewController {
+public class HomeController {
 
     @GetMapping("/reservation")
-    public String world() {
+    public String home() {
         return "reservation";
     }
