+++ src/test/java/roomescape/MissionStepTest.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.Reservation;
 import roomescape.reservation.ReservationDTO;
 
 import static org.assertj.core.api.Assertions.assertThat;
